requirements.txt
redis
pyserial
pillow
